fix: 🐛 check before QuerySet delete

+++ django_softdelete_mixin/query.py
@@ -8,8 +8,19 @@ class SoftDeleteQuerySetMixin:
     @transaction.atomic
     def delete(self, soft=True):
         if soft:
-            collector = SoftDeleteCollector(using=self.db)
-            collector.collect(self)
+            self._not_support_combined_queries("delete")
+            if self.query.is_sliced:
+                raise TypeError("Cannot use 'limit' or 'offset' with delete().")
+            if self.query.distinct or self.query.distinct_fields:
+                raise TypeError("Cannot call delete() after .distinct().")
+            if self._fields is not None:
+                raise TypeError(
+                    "Cannot call delete() after .values() or .values_list()"
+                )
+
+            del_query = self._chain()
+            collector = SoftDeleteCollector(using=del_query.db, origin=self)
+            collector.collect(del_query)
             deleted, _rows_count = collector.delete()
             self._result_cache = None
             return deleted, _rows_count
